requirements.txt
fastai==1.0.51
wandb
tensorboardX==1.6
opencv-python>=3.3.0.10
pillow>=8.0.0
flask
waitress